backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
openpyxl==3.1.2
openai==1.54.0
pypdf==3.17.0
python-dotenv==1.0.0
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.1
httpx==0.25.2
